requirements.txt
gitdb==4.0.5
GitPython==3.1.14
smmap==3.0.5
